requirements.txt
beautifulsoup4==4.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.1
django-polymorphic==0.8.1
django-toolbelt==0.0.1
djangorestframework==3.3.2
feedparser==5.2.1
gunicorn==19.4.5
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
